Impede wrong optimization in the computation of len
We use a #define to set the profile path in the C++ code. Now, the
compiler decides to try to be smart about the fact that it's
a constant so it caches the length of the string, instead of computing
it (as the position of the first \0). But then conda works by
installing in a very long $PREFIX and then doing replacement of that
parefix in the binary with the path of the environment you are
installing in, and filling the rest with zeros.  But since it doesn't
bother to recompute the legth, when you ask for the path it returns
a path of the same length of the original one it was compiled it,
where the new path is padded with zeros.

The solution is to make sure the string is constructed at runtime, by
constricting it from a volatile object.

See:

https://github.com/NNPDF/validphys2/pull/44